src/demo-fetch-weather-data-1/requirements.txt
functions-framework==3.2.0
google-cloud-storage==2.6.0
requests==2.26.0
sqlalchemy
pymysql
cloud-sql-python-connector
